% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifGLMBoost.R
\docType{data}
\name{LearnerClassifGLMBoost}
\alias{LearnerClassifGLMBoost}
\alias{mlr_learners_classif.glmboost}
\title{Boosted Generalized Linear Classification Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerClassif]{mlr3::LearnerClassif}.}
\usage{
LearnerClassifGLMBoost
}
\description{
A \link[mlr3:LearnerClassif]{mlr3::LearnerClassif} for a classification glmboost implemented in \code{\link[mboost:glmboost]{mboost::glmboost()}} in package \CRANpkg{mboost}.
}
\references{
Peter Buhlmann and Bin Yu (2003)
Boosting with the L2 Loss: Regression and Classification
Journal of the American Statistical Association
\url{https://doi.org/10.1198/016214503000125}
}
\keyword{datasets}
