% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrGLMBoost.R
\docType{data}
\name{LearnerRegrGLMBoost}
\alias{LearnerRegrGLMBoost}
\alias{mlr_learners_regr.glmboost}
\title{Boosted Generalized Linear Regression Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrGLMBoost
}
\description{
A \link[mlr3:LearnerRegr]{mlr3::LearnerRegr} for a regression glmboost implemented in \code{\link[mboost:glmboost]{mboost::glmboost()}} in package \CRANpkg{mboost}.
}
\references{
Peter Buhlmann and Bin Yu (2003)
Boosting with the L2 Loss: Regression and Classification
Journal of the American Statistical Association
\url{https://doi.org/10.1198/016214503000125}
}
\keyword{datasets}
