% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learner.mboost-package}
\alias{mlr3learner.mboost}
\alias{mlr3learner.mboost-package}
\title{mlr3learner.mboost: gamboost learner from mboost package for mlr3}
\description{
gamboost learner from mboost package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
