% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learnermboost-package}
\alias{mlr3learnermboost}
\alias{mlr3learnermboost-package}
\title{mlr3learnermboost: gamboost learner from mboost package for mlr3}
\description{
gamboost learner from mboost package for mlr3.
}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

}
