% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifgamboost.R
\docType{data}
\name{LearnerClassifgamboost}
\alias{LearnerClassifgamboost}
\alias{mlr_learners_classif.gamboost}
\title{Classification gamboost Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifgamboost
}
\description{
A \link{LearnerClassif} for a classification gamboost implemented in mboost::gamboost()] in package \CRANpkg{mboost}.
}
\references{
Peter Buhlmann and Bin Yu (2003)
Boosting with the L2 Loss: Regression and Classification
Journal of the American Statistical Association
\url{https://doi.org/10.1198/016214503000125}
}
\keyword{datasets}
