% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrGAMBoost.R
\docType{data}
\name{LearnerRegrGAMBoost}
\alias{LearnerRegrGAMBoost}
\alias{mlr_learners_regr.gamboost}
\title{Boosted Generalized Additive Regression Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link[mlr3:LearnerRegr]{mlr3::LearnerRegr}.}
\usage{
LearnerRegrGAMBoost
}
\description{
A \link{LearnerRegr} for a regression gamboost implemented in \code{\link[mboost:gamboost]{mboost::gamboost()}} in package \CRANpkg{mboost}.
}
\references{
Peter Buhlmann and Bin Yu (2003)
Boosting with the L2 Loss: Regression and Classification
Journal of the American Statistical Association
\url{https://doi.org/10.1198/016214503000125}
}
\keyword{datasets}
