% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.mboost-package}
\alias{mlr3learners.mboost}
\alias{mlr3learners.mboost-package}
\title{mlr3learners.mboost: gamboost learner from mboost package for 'mlr3'}
\description{
gamboost learner from mboost package for
    mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners/mlr3learners.mboost}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de}

Authors:
\itemize{
  \item Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})
}

}
